+++ src/App/navigation/routes/authorizedStackRoutes.ts
@@ -9,3 +9,4 @@ export const authorizedStackRoutes = [
 		},
 	},
 ];
+
