+++ src/App/navigation/routes/unauthorizedStackRoutes.ts
@@ -1,7 +1,5 @@
 import { EmailAuthScreen } from "@screens/EmailAuthScreen";
 import { PhoneAuthScreen } from "@screens/PhoneAuthScreen";
-import { NativeStackScreenProps } from "react-native-screens/native-stack";
-import { FirebaseAuthTypes } from "@react-native-firebase/auth";
 import { OTPVerification } from "@screens/OTPVerification/ui/OTPVerification.tsx";
 
 export const unauthorizedStackRoutes = [
@@ -29,12 +27,5 @@ export const unauthorizedStackRoutes = [
 
 ];
 
-type UnauthorizedStackParamList = {
-	EmailAuthScreen: undefined;
-	PhoneAuthScreen: undefined;
-	OTPVerifyScreen: { confirmation?: FirebaseAuthTypes.ConfirmationResult };
-}
-
-export type UnauthorizedStackRoutesProps = NativeStackScreenProps<UnauthorizedStackParamList, "OTPVerifyScreen">;
