+++ src/shared/api/index.ts
@@ -1,5 +1,6 @@
 export type {UserSchema} from "./config/userSchema/userSchema.ts";
 export type { Nullable } from "./config/common/common.ts";
-export type {RootStackParamList} from "./config/common/navigationTypes/routesType.ts";
+export type {RootStackParamList, UnauthorizedStackRoutesProps} from "./config/common/navigationTypes/routesType.ts";
+
 export { FIRESTORE_DB, FIREBASE_APP, FIREBASE_AUTH, FIREBASE } from "@shared/api/storage/storage.ts";
